Able to run both minibatchers and fit now
It's getting 98% accuracy after only 2 iterations, both for training
and for "validation".

Next step is to run evaluate and predict.  I put validation in quotes,
because it's being evaluated on how well it can determine which digit
is drawn given both a color and a greyscale image of the digit.
A more interesting test will be to see how well it does given only
1 or other other, having been trained on both. I guess to do this,
we'll need to cut one of the inputs out of the model?  But not
how we should load the weights in that case... maybe we just freeze
them for that input?